my-bot/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.29.0
attrs==26.1.0
certifi==2026.4.22
distro==1.9.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.13
jiter==0.14.0
magic-filter==1.0.12
multidict==6.7.1
openai==1.54.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
